valTrackerPY/requirements.txt.txt
requests 
beautifulsoup4 
python-dotenv
